+++ src/main/java/com/TominoCZ/FBP/renderer/FBPRenderer.java
@@ -1,4 +1,4 @@
-package com.TominoCZ.FBP.util;
+package com.TominoCZ.FBP.renderer;
 
 import com.TominoCZ.FBP.FBP;
 import com.TominoCZ.FBP.vector.FBPVector3d;
@@ -7,7 +7,6 @@ import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.texture.TextureMap;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
@@ -17,17 +16,12 @@ import net.minecraft.util.math.Vec3d;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.util.vector.Vector3f;
 
-public class FBPRenderUtil {
+public class FBPRenderer {
 
 
-	public static void renderCubeShaded_S(BufferBuilder buf, Vec2f[] par, float f5, float f6, float f7, double scale, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a, boolean cartoon) {
-		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
-		buf.begin(GL11.GL_QUADS, FBP.POSITION_TEX_COLOR_LMAP_NORMAL);
-
+	public static void renderCubeShaded_S(BufferBuilder buf, Vec2f[] par, float f5, float f6, float f7, double scale, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a) {
 		// switch to vertex format that supports normals
 		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
 		buf.begin(GL11.GL_QUADS, FBP.POSITION_TEX_COLOR_LMAP_NORMAL);
 
 		// some GL commands
@@ -37,28 +31,20 @@ public class FBPRenderUtil {
 		// render particle
 		buf.setTranslation(f5, f6, f7);
 
-		putCube_S(buf, par, scale, rotVec, j, k, r, g, b, a, FBP.cartoonMode);
+		putCube_S(buf, par, scale, rotVec, j, k, r, g, b, a);
 
 		buf.setTranslation(0, 0, 0);
 
 		// continue with the regular vertex format
 		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
 		buf.begin(GL11.GL_QUADS, DefaultVertexFormats.PARTICLE_POSITION_TEX_COLOR_LMAP);
 
 		RenderHelper.disableStandardItemLighting();
 	}
 
-	public static void renderCubeShaded_WH(BufferBuilder buf, Vec2f[] par, float f5, float f6, float f7, double width, double height, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a, boolean cartoon) {
-
+	public static void renderCubeShaded_WH(BufferBuilder buf, Vec2f[] par, float f5, float f6, float f7, double width, double height, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a) {
 		// switch to vertex format that supports normals
 		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
-		buf.begin(GL11.GL_QUADS, FBP.POSITION_TEX_COLOR_LMAP_NORMAL);
-
-		// switch to vertex format that supports normals
-		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
 		buf.begin(GL11.GL_QUADS, FBP.POSITION_TEX_COLOR_LMAP_NORMAL);
 
 		// some GL commands
@@ -68,19 +54,18 @@ public class FBPRenderUtil {
 		// render particle
 		buf.setTranslation(f5, f6, f7);
 
-		putCube_WH(buf, par, width, height, rotVec, j, k, r, g, b, a, FBP.cartoonMode);
+		putCube_WH(buf, par, width, height, rotVec, j, k, r, g, b, a);
 
 		buf.setTranslation(0, 0, 0);
 
 		// continue with the regular vertex format
 		Tessellator.getInstance().draw();
-		Minecraft.getMinecraft().getRenderManager().renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
 		buf.begin(GL11.GL_QUADS, DefaultVertexFormats.PARTICLE_POSITION_TEX_COLOR_LMAP);
 
 		RenderHelper.disableStandardItemLighting();
 	}
 
-	static void putCube_S(BufferBuilder worldRendererIn, Vec2f[] par, double scale, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a, boolean cartoon) {
+	static void putCube_S(BufferBuilder worldRendererIn, Vec2f[] par, double scale, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a) {
 		float radsX = (float) Math.toRadians(rotVec.x);
 		float radsY = (float) Math.toRadians(rotVec.y);
 		float radsZ = (float) Math.toRadians(rotVec.z);
@@ -98,21 +83,15 @@ public class FBPRenderUtil {
 
 			Vec3d normal = rotatef_d(FBP.CUBE_NORMALS[i / 4], radsX, radsY, radsZ);
 
-			if (!cartoon) {
-				addVt_S(worldRendererIn, scale, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v2, par[1].x, par[1].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v3, par[2].x, par[2].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v4, par[3].x, par[3].y, j, k, r, g, b, a, normal);
-			} else {
-				addVt_S(worldRendererIn, scale, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v2, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v3, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_S(worldRendererIn, scale, v4, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-			}
+			addVt_S(worldRendererIn, scale, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
+			addVt_S(worldRendererIn, scale, v2, par[1].x, par[1].y, j, k, r, g, b, a, normal);
+			addVt_S(worldRendererIn, scale, v3, par[2].x, par[2].y, j, k, r, g, b, a, normal);
+			addVt_S(worldRendererIn, scale, v4, par[3].x, par[3].y, j, k, r, g, b, a, normal);
+
 		}
 	}
 
-	static void putCube_WH(BufferBuilder worldRendererIn, Vec2f[] par, double width, double height, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a, boolean cartoon) {
+	static void putCube_WH(BufferBuilder worldRendererIn, Vec2f[] par, double width, double height, FBPVector3d rotVec, int j, int k, float r, float g, float b, float a) {
 		float radsX = (float) Math.toRadians(rotVec.x);
 		float radsY = (float) Math.toRadians(rotVec.y);
 		float radsZ = (float) Math.toRadians(rotVec.z);
@@ -130,17 +109,10 @@ public class FBPRenderUtil {
 
 			Vec3d normal = rotatef_d(FBP.CUBE_NORMALS[i / 4], radsX, radsY, radsZ);
 
-			if (!cartoon) {
-				addVt_WH(worldRendererIn, width, height, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v2, par[1].x, par[1].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v3, par[2].x, par[2].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v4, par[3].x, par[3].y, j, k, r, g, b, a, normal);
-			} else {
-				addVt_WH(worldRendererIn, width, height, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v2, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v3, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-				addVt_WH(worldRendererIn, width, height, v4, par[0].x, par[0].y, j, k, r, g, b, a, normal);
-			}
+			addVt_WH(worldRendererIn, width, height, v1, par[0].x, par[0].y, j, k, r, g, b, a, normal);
+			addVt_WH(worldRendererIn, width, height, v2, par[1].x, par[1].y, j, k, r, g, b, a, normal);
+			addVt_WH(worldRendererIn, width, height, v3, par[2].x, par[2].y, j, k, r, g, b, a, normal);
+			addVt_WH(worldRendererIn, width, height, v4, par[3].x, par[3].y, j, k, r, g, b, a, normal);
 		}
 	}